feat(cce): support dss in node and node pool
