+++ docs/resources/cce_node_pool.md
@@ -273,6 +273,9 @@ The `root_volume` block supports:
 * `kms_key_id` - (Optional, String, ForceNew) Specifies the KMS key ID. This is used to encrypt the volume.
   Changing this parameter will create a new resource.
 
+* `dss_pool_id` - (Optional, String, ForceNew) Specifies the DSS pool ID. This field is used only for dedicated storage.
+  Changing this parameter will create a new resource.
+
 The `data_volumes` block supports:
 
 * `size` - (Required, Int, ForceNew) Specifies the disk size in GB. Changing this parameter will create a new resource.
@@ -283,6 +286,9 @@ The `data_volumes` block supports:
   Changing this parameter will create a new resource.
 
 * `kms_key_id` - (Optional, String, ForceNew) Specifies the KMS key ID. This is used to encrypt the volume.
+  Changing this parameter will create a new resource.
+
+* `dss_pool_id` - (Optional, String, ForceNew) Specifies the DSS pool ID. This field is used only for dedicated storage.
   Changing this parameter will create a new resource.
 
   -> You need to create an agency (EVSAccessKMS) when disk encryption is used in the current project for the first time ever.
