+++ huaweicloud/services/cce/resource_huaweicloud_cce_node.go
@@ -88,95 +88,8 @@ func ResourceNode() *schema.Resource {
 				Optional:  true,
 				Sensitive: true,
 			},
-			"root_volume": {
-				Type:     schema.TypeList,
-				Required: true,
-				ForceNew: true,
-				MaxItems: 1,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"size": {
-							Type:     schema.TypeInt,
-							Required: true,
-							ForceNew: true,
-						},
-						"volumetype": {
-							Type:     schema.TypeString,
-							Required: true,
-							ForceNew: true,
-						},
-						"hw_passthrough": {
-							Type:        schema.TypeBool,
-							Optional:    true,
-							ForceNew:    true,
-							Description: "schema: Internal",
-						},
-						"extend_param": {
-							Type:       schema.TypeString,
-							Optional:   true,
-							ForceNew:   true,
-							Deprecated: "use extend_params instead",
-						},
-						"extend_params": {
-							Type:     schema.TypeMap,
-							Optional: true,
-							ForceNew: true,
-							Computed: true,
-							Elem:     &schema.Schema{Type: schema.TypeString},
-						},
-						"kms_key_id": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-							ForceNew: true,
-						},
-					},
-				},
-			},
-			"data_volumes": {
-				Type:     schema.TypeList,
-				Required: true,
-				ForceNew: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"size": {
-							Type:     schema.TypeInt,
-							Required: true,
-							ForceNew: true,
-						},
-						"volumetype": {
-							Type:     schema.TypeString,
-							Required: true,
-							ForceNew: true,
-						},
-						"hw_passthrough": {
-							Type:        schema.TypeBool,
-							Optional:    true,
-							ForceNew:    true,
-							Description: "schema: Internal",
-						},
-						"extend_param": {
-							Type:       schema.TypeString,
-							Optional:   true,
-							ForceNew:   true,
-							Deprecated: "use extend_params instead",
-						},
-						"extend_params": {
-							Type:     schema.TypeMap,
-							Optional: true,
-							ForceNew: true,
-							Computed: true,
-							Elem:     &schema.Schema{Type: schema.TypeString},
-						},
-						"kms_key_id": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-							ForceNew: true,
-						},
-					},
-				},
-			},
+			"root_volume":  resourceNodeRootVolume(),
+			"data_volumes": resourceNodeDataVolume(),
 			"storage": {
 				Type:     schema.TypeList,
 				Optional: true,
@@ -535,50 +448,6 @@ func buildResourceNodeTags(d *schema.ResourceData) []tags.ResourceTag {
 	return utils.ExpandResourceTags(tagRaw)
 }
 
-func buildResourceNodeRootVolume(d *schema.ResourceData) nodes.VolumeSpec {
-	var root nodes.VolumeSpec
-	volumeRaw := d.Get("root_volume").([]interface{})
-	if len(volumeRaw) == 1 {
-		rawMap := volumeRaw[0].(map[string]interface{})
-		root.Size = rawMap["size"].(int)
-		root.VolumeType = rawMap["volumetype"].(string)
-		root.HwPassthrough = rawMap["hw_passthrough"].(bool)
-		root.ExtendParam = rawMap["extend_params"].(map[string]interface{})
-
-		if rawMap["kms_key_id"].(string) != "" {
-			metadata := nodes.VolumeMetadata{
-				SystemEncrypted: "1",
-				SystemCmkid:     rawMap["kms_key_id"].(string),
-			}
-			root.Metadata = &metadata
-		}
-	}
-
-	return root
-}
-
-func buildResourceNodeDataVolume(d *schema.ResourceData) []nodes.VolumeSpec {
-	volumeRaw := d.Get("data_volumes").([]interface{})
-	volumes := make([]nodes.VolumeSpec, len(volumeRaw))
-	for i, raw := range volumeRaw {
-		rawMap := raw.(map[string]interface{})
-		volumes[i] = nodes.VolumeSpec{
-			Size:          rawMap["size"].(int),
-			VolumeType:    rawMap["volumetype"].(string),
-			HwPassthrough: rawMap["hw_passthrough"].(bool),
-			ExtendParam:   rawMap["extend_params"].(map[string]interface{}),
-		}
-		if rawMap["kms_key_id"].(string) != "" {
-			metadata := nodes.VolumeMetadata{
-				SystemEncrypted: "1",
-				SystemCmkid:     rawMap["kms_key_id"].(string),
-			}
-			volumes[i].Metadata = &metadata
-		}
-	}
-	return volumes
-}
-
 func buildResourceNodeTaint(d *schema.ResourceData) []nodes.TaintSpec {
 	taintRaw := d.Get("taints").([]interface{})
 	taints := make([]nodes.TaintSpec, len(taintRaw))
@@ -926,46 +795,6 @@ func resourceNodeRead(_ context.Context, d *schema.ResourceData, meta interface{
 		return diag.Errorf("error setting CCE Node fields: %s", err)
 	}
 	return nil
-}
-
-func flattenResourceNodeRootVolume(rootVolume nodes.VolumeSpec) []map[string]interface{} {
-	res := []map[string]interface{}{
-		{
-			"size":           rootVolume.Size,
-			"volumetype":     rootVolume.VolumeType,
-			"hw_passthrough": rootVolume.HwPassthrough,
-			"extend_params":  rootVolume.ExtendParam,
-			"extend_param":   "",
-		},
-	}
-	if rootVolume.Metadata != nil {
-		res[0]["kms_key_id"] = rootVolume.Metadata.SystemCmkid
-	}
-
-	return res
-}
-
-func flattenResourceNodeDataVolume(dataVolumes []nodes.VolumeSpec) []map[string]interface{} {
-	if len(dataVolumes) == 0 {
-		return nil
-	}
-
-	res := make([]map[string]interface{}, len(dataVolumes))
-	for i, v := range dataVolumes {
-		res[i] = map[string]interface{}{
-			"size":           v.Size,
-			"volumetype":     v.VolumeType,
-			"hw_passthrough": v.HwPassthrough,
-			"extend_params":  v.ExtendParam,
-			"extend_param":   "",
-		}
-
-		if v.Metadata != nil {
-			res[i]["kms_key_id"] = v.Metadata.SystemCmkid
-		}
-	}
-
-	return res
 }
 
 func resourceNodeUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
