+++ huaweicloud/services/cce/resource_huaweicloud_cce_node_pool.go
@@ -74,101 +74,8 @@ func ResourceNodePool() *schema.Resource {
 				Optional: true,
 				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
-			"root_volume": {
-				Type:     schema.TypeList,
-				Required: true,
-				ForceNew: true,
-				MaxItems: 1,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"size": {
-							Type:     schema.TypeInt,
-							Required: true,
-							ForceNew: true,
-						},
-						"volumetype": {
-							Type:     schema.TypeString,
-							Required: true,
-							ForceNew: true,
-						},
-						"extend_params": {
-							Type:     schema.TypeMap,
-							Optional: true,
-							ForceNew: true,
-							Elem:     &schema.Schema{Type: schema.TypeString},
-						},
-						"kms_key_id": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-							ForceNew: true,
-						},
-
-						// Internal parameters
-						"hw_passthrough": {
-							Type:        schema.TypeBool,
-							Optional:    true,
-							ForceNew:    true,
-							Description: "schema: Internal",
-						},
-
-						// Deprecated parameters
-						"extend_param": {
-							Type:       schema.TypeString,
-							Optional:   true,
-							ForceNew:   true,
-							Deprecated: "use extend_params instead",
-						},
-					},
-				},
-			},
-			"data_volumes": {
-				Type:     schema.TypeList,
-				Required: true,
-				ForceNew: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"size": {
-							Type:     schema.TypeInt,
-							Required: true,
-							ForceNew: true,
-						},
-						"volumetype": {
-							Type:     schema.TypeString,
-							Required: true,
-							ForceNew: true,
-						},
-						"extend_params": {
-							Type:     schema.TypeMap,
-							Optional: true,
-							ForceNew: true,
-							Elem:     &schema.Schema{Type: schema.TypeString},
-						},
-						"kms_key_id": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-							ForceNew: true,
-						},
-
-						// Internal parameters
-						"hw_passthrough": {
-							Type:        schema.TypeBool,
-							Optional:    true,
-							ForceNew:    true,
-							Description: "schema: Internal",
-						},
-
-						// Deprecated parameters
-						"extend_param": {
-							Type:       schema.TypeString,
-							Optional:   true,
-							ForceNew:   true,
-							Deprecated: "use extend_params instead",
-						},
-					},
-				},
-			},
+			"root_volume":  resourceNodeRootVolume(),
+			"data_volumes": resourceNodeDataVolume(),
 			"availability_zone": {
 				Type:     schema.TypeString,
 				Optional: true,
